api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.0.0
openai==1.95.1
faiss-cpu==1.11.0
sentence-transformers==4.1.0
pandas==2.3.0
numpy==2.2.6
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.20